🗑️ chore: Remove unused resume models and refactor related functionality
- Deleted `ParsedResume`, `Resume`, and `TargetJob` models from `resume.models` as they were unused.
- Moved relevant fields and logic from `Resume` to `InterviewSession` in `interview.models`.
- Updated `InterviewSession` with new fields, statuses, and functionality for better integration.
- Refactored admin configurations to align with model changes, including removing `ParsedResumeAdmin` and `ResumeAdmin`.
