+++ backend/resume/admin.py
@@ -1,19 +1,6 @@
 from django.contrib import admin
-from .models import ParsedResume, Resume
 
-@admin.register(Resume)
-class ResumeAdmin(admin.ModelAdmin):
-    list_display = ('id', 'local_path', 'uploaded_at', 'target_job')
-    list_filter = ('uploaded_at', 'target_job')
-    search_fields = ('id', 'target_job')
-    readonly_fields = ('id', 'uploaded_at')
-    
-    def has_add_permission(self, request):
-        # Prevent manual creation through admin - should be uploaded via API
-        return False
-
-@admin.register(ParsedResume)
-class ParsedResumeAdmin(admin.ModelAdmin):
-    list_display = ('id', 'user', 'uploaded_at')
-    list_filter = ('uploaded_at',)
-    search_fields = ('user__email',)
+# NOTE: All models have been moved or removed:
+# - Resume model moved to interview.models.InterviewSession 
+# - ParsedResume and TargetJob models removed as they were unused
+# - InterviewSession is registered in interview/admin.py
